+++ src/OpenWOO/Models/GeografischePositieEntity.php
@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace Yard\OpenWOO\Models;
 
@@ -12,9 +14,12 @@ class GeografischePositieEntity extends AbstractEntity
             return [];
         }
 
+        $longitude = $this->data[self::PREFIX . 'Longitude'] ?? null;
+        $lattitude = $this->data[self::PREFIX . 'Lattitude'] ?? null;
+
         return [
-            'Longitude' => ! empty($this->data[self::PREFIX . 'Longitude']) ? (float) $this->data[self::PREFIX . 'Longitude'] : '',
-            'Lattitude' => ! empty($this->data[self::PREFIX . 'Lattitude']) ? (float) $this->data[self::PREFIX . 'Lattitude'] : '',
+            'Longitude' => ! empty($longitude) && is_numeric($longitude) ? (float) $longitude : '',
+            'Lattitude' => ! empty($lattitude) && is_numeric($lattitude) ? (float) $lattitude : '',
         ];
     }
 }